requirements.txt
aiogram==3.4.1
openai==1.54.4
google-cloud-firestore==2.20.0
aiohttp~=3.9.0
httpx==0.27.2
google-cloud-tasks==2.16.3
